requirements.txt
requests
beautifulsoup4
tldextract
python-whois
